// test/generated/dist/offices/offices.models.ts
export namespace OfficesModels {
// ...
  export const OfficeBookkeepingResponseDtoSchema = z
    .object({
      generalLedgerSystem: z.string(),
      defaultCurrency: OfficeCurrencyResponseDtoSchema.nullish(),
      availableCurrencies: z.array(OfficeCurrencyResponseDtoSchema).readonly(),
      showPaymentInstructions: z.boolean(),
      showCompanyRegistrationNumber: z.boolean(),
      showInvoiceVatLinesInOfficeCurrency: z.boolean(),
      reportInvoicesToHungarianTaxAuthority: z.boolean(),
      bankAccountCurrencyMapping: z.array(OfficeBankAccountCurrencyMappingResponseDtoSchema).readonly(),
      costCenterId: z.string().nullish(),
      minimumOutgoingInvoiceDate: z.iso.datetime({ offset: true }).nullish(),
      minimumOutgoingInvoiceServiceDate: z.iso.datetime({ offset: true }).nullish(),
      minimumIncomingInvoiceDate: z.iso.datetime({ offset: true }).nullish(),
      minimumIncomingInvoiceServiceDate: z.iso.datetime({ offset: true }).nullish(),
      factoringReportingEnabled: z.boolean(),
    })
    .readonly();
// ...
  export const OfficeDocumentSettingsResponseDtoSchema = z
    .object({
      id: z.string(),
      mediaUrl: z.string(),
      language: CommonModels.LanguageEnumSchema,
      documentPart: OfficeDocumentPartEnumSchema,
    })
    .readonly();
// ...
  export const OfficeRepresentingBusinessPartnerResponseDtoSchema = z
    .object({
      id: z.string(),
      name: z.string(),
      types: z.array(z.string()).readonly(),
      email: z.array(z.string()).readonly().nullish(),
      phone: z.array(z.string()).readonly().nullish(),
      addressId: z.string().nullish(),
      address: AddressResponseDTOSchema.nullish(),
    })
    .readonly();
  export type OfficeRepresentingBusinessPartnerResponseDto = z.infer<
    typeof OfficeRepresentingBusinessPartnerResponseDtoSchema
  >;

  /**
   * OfficePaymentTermsResponseDtoSchema
   * @type { object }
   * @property { number } days
   * @property { string } relativeTo
   */
  export const OfficePaymentTermsResponseDtoSchema = z
    .object({ days: z.number(), relativeTo: CommonModels.OfficePaymentTermsDateTypeSchema })
    .readonly();
  export type OfficePaymentTermsResponseDto = z.infer<typeof OfficePaymentTermsResponseDtoSchema>;

  /**
   * OfficeCountryResponseDtoSchema
   * @type { object }
   * @property { string } id
   * @property { string } name
   */
  export const OfficeCountryResponseDtoSchema = z.object({ id: z.string(), name: z.string() }).readonly();
// ...
  export const OfficeBankAccountResponseDtoSchema = z
    .object({
      id: z.string(),
      name: z.string(),
      bankName: z.string(),
      iban: z.string(),
      swiftBic: z.string(),
      useFooterOnInvoice: z.boolean(),
      footer: z.object({ mediaUrl: z.string() }).readonly().nullish(),
      isFactoringBank: z.boolean(),
    })
    .readonly();
// ...
  export const OfficeDetailResponseDtoSchema = z
    .object({
      id: z.string(),
      name: z.string(),
      symbol: z.string().nullable(),
      invoicePrefix: z.number().nullable(),
      autoClosePositionsAfterDays: z.number().nullish(),
      defaultLanguage: z.string().nullable(),
      locale: CommonModels.LocaleEnumSchema.nullish(),
      vatId: z.string().nullable(),
      taxNumber: z.string().nullable(),
      restrictPositionInvolvedParties: z.boolean(),
      showWatermarkOnDocuments: z.boolean(),
      usePartnerMatchCodes: z.boolean(),
      restrictFinancePartnersToRelationship: z.boolean(),
      termsExport: OfficePaymentTermsResponseDtoSchema.nullable(),
      termsImport: OfficePaymentTermsResponseDtoSchema.nullable(),
      country: OfficeCountryResponseDtoSchema.nullable(),
      bankAccounts: z.array(OfficeBankAccountResponseDtoSchema).readonly(),
      bookkeeping: OfficeBookkeepingResponseDtoSchema,
      documentSettings: z.array(OfficeDocumentSettingsResponseDtoSchema).readonly(),
      representingBusinessPartner: OfficeRepresentingBusinessPartnerResponseDtoSchema.nullable(),
      hasInttraCredentials: z.boolean().nullish(),
    })
    .readonly();
// ...
  export const UpdateOfficePaymentTermsRequestSchema = z
    .object({ days: z.number(), relativeTo: CommonModels.OfficePaymentTermsDateTypeSchema })
    .readonly();
  export type UpdateOfficePaymentTermsRequest = z.infer<typeof UpdateOfficePaymentTermsRequestSchema>;

  /**
   * UpdateOfficeBankAccountCurrencyMappingSchema
   * @type { object }
   * @property { string } currency
   * @property { string } bankAccountId
   */
  export const UpdateOfficeBankAccountCurrencyMappingSchema = z
    .object({ currency: z.string(), bankAccountId: z.string() })
    .readonly();
// ...
  export const UpdateOfficeRequestSchema = z
    .object({
      name: z.string().describe("Office Name"),
      symbol: z.string().describe("Office Symbol"),
      invoicePrefix: z.number().describe("Office Invoice Prefix"),
      autoClosePositionsAfterDays: z.number(),
      defaultLanguage: z.string().describe("Office Default Language"),
      locale: CommonModels.LocaleEnumSchema,
      vatId: z.string().describe("Office Vat Id"),
      taxNumber: z.string().describe("Office Tax Number"),
      termsExport: UpdateOfficePaymentTermsRequestSchema.describe("Office Terms Export"),
      termsImport: UpdateOfficePaymentTermsRequestSchema.describe("Office Terms Import"),
      showPaymentInstructions: z.boolean().describe("Office Show Payment Instructions"),
      showCompanyRegistrationNumber: z.boolean(),
      reportInvoicesToHungarianTaxAuthority: z.boolean(),
      restrictPositionInvolvedParties: z.boolean(),
      showWatermarkOnDocuments: z.boolean(),
      showInvoiceVatLinesInOfficeCurrency: z.boolean(),
      usePartnerMatchCodes: z.boolean(),
      restrictFinancePartnersToRelationship: z.boolean(),
      costCenterId: z.string(),
      minimumOutgoingInvoiceDate: z.iso.datetime({ offset: true }),
      minimumOutgoingInvoiceServiceDate: z.iso.datetime({ offset: true }),
      minimumIncomingInvoiceDate: z.iso.datetime({ offset: true }),
      minimumIncomingInvoiceServiceDate: z.iso.datetime({ offset: true }),
      countryId: z.string().describe("Office country id"),
      defaultCurrencyId: z.string().describe("Office default currency id"),
      representingBusinessPartnerId: z.string().describe("Representing business partner id"),
      availableCurrencyIds: z.array(z.string()).readonly().describe("Office availableCurrencyIds"),
      generalLedgerSystem: z.string().describe("Office general ledger system"),
      factoringReportingEnabled: z.boolean(),
      bankAccountCurrencyMapping: z
        .array(UpdateOfficeBankAccountCurrencyMappingSchema)
        .readonly()
        .describe("Office bank account currency mapping"),
    })
    .readonly();
// ...
  export const OfficeListItemResponseSchema = z
    .object({ id: z.string(), name: z.string(), numberOfEmployees: z.number() })
    .readonly();
  export type OfficeListItemResponse = z.infer<typeof OfficeListItemResponseSchema>;

  /**
   * OfficeFilterDtoSchema
   * @type { object }
   * @property { string } name Name
   * @property { string } search
   */
  export const OfficeFilterDtoSchema = z.object({ name: z.string().describe("Name"), search: z.string() }).readonly();
  export type OfficeFilterDto = z.infer<typeof OfficeFilterDtoSchema>;

  /**
   * OfficeLabelFilterDtoSchema
   * @type { object }
   * @property { string } search
   */
  export const OfficeLabelFilterDtoSchema = z.object({ search: z.string() }).readonly();
  export type OfficeLabelFilterDto = z.infer<typeof OfficeLabelFilterDtoSchema>;

  /**
   * CreateOfficeRequestSchema
   * @type { object }
   * @property { string } name
   */
  export const CreateOfficeRequestSchema = z.object({ name: z.string() }).readonly();
// ...
  export const OfficesPaginateOrderParamEnumSchema = z.enum(["name"]);
// ...
  export const OfficesPaginateLabelsOrderParamEnumSchema = z.enum(["name"]);
// ...
}